+++ xarray/__init__.py
@@ -1,3 +1,4 @@
+# flake8: noqa
 from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function